implement minimal templating for output js
In substitution.json, it is possible to refer to a template placeholder
within a specific javascript file (.js extentsion). The format is the following: [filename$placeHolder]. Example: with filename=task.js and placeholder=test, it should be [task.js$test]
In addition, a new optional field can be added to the substitutions: "keepArray". For now, it behaves like required in html forms. That is,
it's value isn't used anywhere, but we check if the array needs to be substituted literally.